Separate out Test/Lib
There are some helper modules that are convenient for other projects to use, e.g. `Signet.Test.Client` which acts as a mock for a variety of common function calls. We want these modules to be avilable, and thus they cannot simply be in test/support since those are not included at all in downstream projects. This patch adds them to `test/lib`, but also adds that as an elixirc path when in Mix.env=test, which should compile these for downstream deps.