Add Base Fee & Priority Fee Support
This patch starts to add support to building a transaction based on base fee and priority fee. These are the first changes to start to make Signet compliant with EIP-1559, but we still currently add the fees together and wrap them in a Legacy transaction. This will more or less give the same functionality as users should specify `priority_fee` and let the base fee be handled by the protocol.

Note: this isn't great since we add a 20% buffer to the base fee that _becomes_ a priority fee. We'll fix this in future upgrades.