Handle Non-EIP-155 Signatures
According to [EIP-155](https://eips.ethereum.org/EIPS/eip-155), signatures for Ethereum transactions or messages use a recovery bit that is `35 + chain_id * 2 + rec_id`, which we took to mean an odd recovery bit would mean the original rec_id was 0, and an even would mean the original rec_id was 1. But when hardware wallets and other people sign and don't follow EIP-155, then 0 should still imply a rec_id of 0, and 1 should still imply a rec_id of 1, and thus we add explicit support for that in this patch.