Adjust RPC Call Return Value (1.0.0-alpha6)
Previously we had `send_rpc` return a tuple of the form `{:error, String.t()}`, which has problematic since in many cases that data was less useful than the error returned from the JSON-RPC node (e.g. the error code got put into that string). Here, we fix this by making this call always return `{:error, %{code: code, message: message}}`, which is easier for clients to parse. Additionally parsed information, such as `%{code: 3, message: "execution reverted", revert: <<1,2,3>>, error_abi: "Cool()", error_params: []}` can expand this struct, but it will always have code and message on it. This should be a boon for callers. Note: this is a breaking change and could break earlier clients.

Bump to 1.0.0-alpha6