Allow Wide V2 Signatures
Due to EIP-155, some signatures come in wide (e.g. with a very large recovery "bit"), this is esp. true for nets like Sepolia with chain_id=11155111. Signatures should be passed in in a nice parsed way (r, s, and v), but if they are binary-packed, then EIP-155 means that large nets will overflow 1-byte (originally allocated for `v`). Here, we expand the signature to allow wider-`v`s.

Bump to 1.0.0-beta3