Add `trace_callMany` Support [1.0.0-beta5]
This patch adds support for [trace_callMany](https://docs.infura.io/api/networks/ethereum/json-rpc-methods/trace-methods/trace_callmany), which is supported by some network providers to speculative execute multiple transactions in a single call (e.g. similar to `eth_call`). This can be used in contexts where you may want to measure the gas or effect of a transaction, but a single `eth_call` may not be able to achieve the required output.