Improve Hex Library
We've added a bunch of beneficial helpers to `Signet.Hex` instead of the ones from `Signet.Util`. The biggest change is the addition of a helper sigil:

```elixir
x = ~h[0x112233]
```

That will parse the hex string at compile time, which is useful for inlined constants and crashing at compile-time instead of runtime on those values.