Fix ABI-only Encoding
This patch fixes ABI-only generators, e.g. when you _just_ have the ABI, and not the full Solidity output. This allows you to create stub contracts useful for calling, but that don't have "bytecode" or "constructor." Previously we _mostly_ had this working, except the constructor was still being created which was then causing a compilation error since it referenced `bytecode/0` that wasn't being defined.

Additionally, we have a lot of work to do on name overlaps. E.g. in this case there were two functions "abc()" and "ABC()" which Solidity treats as different but end up the same name in our generated code. We fix that, but honestly, we need a bigger system for checking for dups that goes deeper than this patch does.