Filter Improvements
This patch makes two changes to `Signet.Filter`.

First, we accept `rpc_opts` that we pass to our RPC calls. E.g. you can now specity an `ethereum_node: "..."` and that will be used in the calls from `Signet.Filter`.

Secondly, we use our passed in `event`-types to build our topics to filter on (this is obvious, right?). That said, there's a weird incongruity that we need to work on. Specifically, when you try to decode a topic, the first indexed parameter is the "topic"-itself (e.g. the hash of the event signature). This is annoying, so we recnetly started adding a first parameter "_topic" so that it decodes correctly. We should probably change ABI so that `decode_event` properly handles popping that first topic, but for now, we just align this module with that so that it properly can decode events. Again, this is a bit of a hack, but the change is way more complex and reaches outside of this repo.