Sleuth Binaries
This patch adds the ability to _not decode_ binaries coming back from sleuth (e.g. addresses and bytes) when returning a sleuth query. This is because they are generally non-serializable for JSON in binary form, and it may be convenient to simply have them left as hex strings when getting the sleuth result.

We also fix up our sleuth unit tests to be significantly more legible and maintainable.

Bump to delta5