Trace Reverts
This patch adds the ability to easily trace reverts on `call_trx` calls, e.g. so you can easily see the revert reasons. This simply has the ability to try and trace a call after it reverts (by setting `trace_reverts: true`) and then an error that reverted will include a `trace: Trace{}` field that can be observed.

Bump to delta7