Add FFI Integration
This patch adds the ability to call into FFI "foreign-function interfaces," which are hooks back into the EVM's calling context environment. From the perspective of the EVM/Solidity code, this is the same as calling a standard Ethereum precompile (e.g. these functions: https://www.evm.codes/precompiled), but overall, this means we can allow hooks from EVM-VM code, e.g. to fetch offline price data. This lightly violates the concept of being pure, but overall, it's up to the discretion of the author to add hooks or not into code.

Note: we change the signature of `Signet.VM.exec` and `Signet.VM.exec_call` to use `exec_opts`, which allows FFIs, and thus, we change some of our generated code, as well, to match this new format.