Add Trace Serializers
This patch adds the opposite of `trace.deserialize()` which allows us to
re-serialize a JSON (be it a call trace, a trace trx or a debug trace).
These are useful for generally printing or showing traces, which can be
difficult due to the amount of Elixir-binaries that usually show up.