Fix Transaction Encoding/Decoding
This patch "fixes" transaction encoding and decoding, specifically
focused on trimming signature r and signature s of prefixed zeros, which
seem to be against the EIP-2718 spec in ways I don't understand. That
said, this fixes most clients that receive an error in such cases and
the test cases now work with "cast dt 0x..." to ensure they match
expectations of the `alloy` ethereum client.

Additionally, we fix access lists, which were never used but incorrect
previously, as well.